// src/schemas/schemas.ts
import { z } from 'zod'

export const serieASchema = z.object({
  giornata: z.number(),
  squadraHome: z.string(),
  squadraAway: z.string(),
})

export const giornataSchema = z.object({
  idCalendario: z.number(),
  idTorneo: z.number(),
  giornata: z.number(),
  giornataSerieA: z.number(),
  isGiocata: z.boolean(),
  isSovrapposta: z.boolean(),
  isRecupero: z.boolean(),
  data: z.string().optional(),
  dataFine: z.string().optional(),
  girone: z.union([z.string(), z.number(), z.null()]),
  partite: z
    .object({
      idPartita: z.number(),
      idHome: z.number().nullable(),
      squadraHome: z.string().nullable().optional(),
      fotoHome: z.string().nullable().optional(),
      magliaHome: z.string().nullable().optional(),
      multaHome: z.boolean(),
      golHome: z.number().nullable(),
      idAway: z.number().nullable(),
      squadraAway: z.string().nullable().optional(),
      fotoAway: z.string().nullable().optional(),
      magliaAway: z.string().nullable().optional(),
      multaAway: z.boolean(),
      golAway: z.number().nullable(),
      isFattoreHome: z.boolean(),
    })
    .array(),
  Torneo: z.string(),
  Descrizione: z.string(),
  Title: z.string(),
  SubTitle: z.string(),
  SerieA: z.array(serieASchema).optional(),
})

export const torneiSchema = z.object({
  idTorneo: z.number(),
  nome: z.string(),
  gruppoFase: z.string().nullable(),
})

const utentePartitaSchema = z.object({
  nomeSquadra: z.string().nullable(),
  foto: z.string().nullable(),
  maglia: z.string().nullable(),
})

export const partitaSchema = z.object({
  idPartita: z.number(),
  idSquadraH: z.number().nullable(),
  idSquadraA: z.number().nullable(),
  hasMultaH: z.boolean(),
  hasMultaA: z.boolean(),
  golH: z.number().nullable(),
  golA: z.number().nullable(),
  fattoreCasalingo: z.boolean(),
  Utenti_Partite_idSquadraHToUtenti: utentePartitaSchema.nullable(),
  Utenti_Partite_idSquadraAToUtenti: utentePartitaSchema.nullable(),
})

export const calendarioPartiteSchema = z.object({
  idCalendario: z.number(),
  giornata: z.number(),
  giornataSerieA: z.number(),
  hasGiocata: z.boolean(),
  hasSovrapposta: z.boolean(),
  hasDaRecuperare: z.boolean(),
  data: z.date().nullable(),
  dataFine: z.date().nullable(),
  girone: z.union([z.string(), z.number()]).nullable(),
  Tornei: torneiSchema,
  Partite: z.array(partitaSchema),
})

export const calendarioSchema = z.object({
  id: z.number().default(0),
  idTorneo: z.number().default(1),
  nome: z.string().default(''),
  gruppoFase: z.string().nullable().default(null),
  giornata: z.number().default(0),
  giornataSerieA: z.number().default(0),
  isGiocata: z.boolean().default(false),
  isSovrapposta: z.boolean().default(false),
  isRecupero: z.boolean().default(false),
  data: z.string().optional().default(''),
  dataFine: z.string().optional().default(''),
  girone: z.number().nullable().default(null),
  isSelected: z.boolean().default(false),
})

export const classificaSchema = z.object({
  idSquadra: z.number(),
  squadra: z.string(),
  foto: z.string().nullable(),
  punti: z.number(),
  vinte: z.number(),
  pareggi: z.number(),
  perse: z.number(),
  golFatti: z.number(),
  golSubiti: z.number(),
  differenzaReti: z.number(),
  giocate: z.number(),
  fantapunti: z.number(),
})